+++ R/oa_fetch.R
@@ -7,7 +7,7 @@
 oa_entities <- function() {
   c(
     "works", "authors", "institutions", "concepts",
-    "funders", "sources", "publishers"
+    "funders", "sources", "publishers", "topics"
   )
 }
 
@@ -461,7 +461,7 @@ get_next_page <- function(paging, i, res = NULL) {
 #' See more at <https://docs.openalex.org/how-to-use-the-api/get-single-entities#the-openalex-id>.
 #' @param entity Character. Scholarly entity of the search.
 #' The argument can be one of
-#' c("works", "authors", "institutions", "concepts", "funders", "sources", "publishers").
+#' c("works", "authors", "institutions", "concepts", "funders", "sources", "publishers", "topics").
 #' If not provided, `entity` is guessed from `identifier`.
 #' @param options List. Additional parameters to add in the query. For example:
 #'